Added IUPHAR database retriever
So - this is a long one. The website gives only some data in .csv form,
and ofc we need some data that they do not digest for us.
They provide a REST API that you need to call 800*3 times to get al the
data that we need. It also crashes randomly.

So the last thing to do is to download their db dump and load it. But it
is in the wrong format, even for SQLite!
You either load an instance of POSTGRESql, load it and re-dump it into a
dump that in SQLite-readable, or do it like I did. I look for the raw
data in the dump and load that. I hope I do it right...