Cleared up and shortenede ion channels parser
The ion channel parser was very dirty from the conversion that happened
a while ago where we used one col per ion and moved to just one "carried
solute" column. We were doing a lot of useless transformations to go
back and forth. This fixes the issue, as all the back-and-forth is gone.

I also added in a lot of comments on the shape of the data, and comments
on why I do the stuff I do. It sholud make the code clearer to any
future me that reads it.