Added 'origin' table
Added the expression tissue table, with where the various genes are
expressed. The data is sourced from the protein atlas.

We basically select just the columns that we want, collapse and explode
some of them, like the different 'skin' levels. We don't need them for
the use we have of the DB.