fix: dropped empty string as a carried solute in one gene
For reasons unknown, ENSG00000074621 had "" as a carried solute, and
this avoided filtering. Since we don't know a priori if cases like this
should be handled by dropping the row or replacing with NULL, I handled
it manually in the very first post build hook. In this case, the entry
is dropped (as the slc is annotated correctly in other rows).